requirements.txt
certifi==2023.7.22
charset-normalizer==3.3.0
dataclasses-json==0.5.14
httplib2==0.22.0
idna==3.4
instagrapi==2.0.0
marshmallow==3.20.1
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==23.2
Pillow==10.0.1
pycryptodomex==3.18.0
pydantic==1.10.9
pyparsing==3.1.1
PySocks==1.7.1
python-dotenv==1.0.0
python-facebook-api==0.17.1
requests==2.31.0
requests-oauthlib==1.3.1
simplejson==3.19.2
six==1.16.0
typing-inspect==0.9.0
typing_extensions==4.8.0
urllib3==2.0.6
